+++ src/components/FeaturePanel/ObjectsAround.tsx
@@ -4,10 +4,9 @@ import Router from 'next/router';
 import { fetchAroundFeature } from '../../services/osmApi';
 import { useFeatureContext } from '../utils/FeatureContext';
 import { Feature } from '../../services/types';
-import { getOsmappLink, getUrlOsmId } from '../../services/helpers';
+import { FetchError, getOsmappLink, getUrlOsmId } from "../../services/helpers";
 import Maki from '../utils/Maki';
 import { t } from '../../services/intl';
-import { FetchError } from '../../services/fetch';
 import { trimText, useMobileMode } from '../helpers';
 import { getLabel } from '../../helpers/featureLabel';
 import { useUserThemeContext } from '../../helpers/theme';
