+++ src/services/fetch.ts
@@ -1,21 +1,7 @@
-import fetch from 'isomorphic-unfetch';
-import { getCache, getKey, writeCacheSafe } from './fetchCache';
-import { isBrowser } from '../components/helpers';
-
-export class FetchError extends Error {
-  constructor(
-    public message: string = '',
-    public code: string,
-    public data: string,
-  ) {
-    super();
-  }
-
-  toString() {
-    const suffix = this.data && ` Data: ${this.data.substring(0, 1000)}`;
-    return `Fetch: ${this.message}${suffix}`;
-  }
-}
+import fetch from "isomorphic-unfetch";
+import { getCache, getKey, writeCacheSafe } from "./fetchCache";
+import { isBrowser } from "../components/helpers";
+import { FetchError } from "./helpers";
 
 // TODO cancel request in map.on('click', ...)
 const abortableQueues: Record<string, AbortController> = {};
