+++ src/services/helpers.ts
@@ -1,10 +1,10 @@
-import * as xml2js from 'isomorphic-xml2js';
-import fetch from 'isomorphic-unfetch';
-import { isServer, isString } from '../components/helpers';
-import { Feature, Position } from './types';
-import { join, roundedToDegUrl } from '../utils';
-import { PROJECT_URL } from './project';
-import { getIdFromShortener, getShortenerSlug } from './shortener';
+import * as xml2js from "isomorphic-xml2js";
+import fetch from "isomorphic-unfetch";
+import { isServer, isString } from "../components/helpers";
+import { Feature, Position } from "./types";
+import { join, roundedToDegUrl } from "../utils";
+import { PROJECT_URL } from "./project";
+import { getIdFromShortener, getShortenerSlug } from "./shortener";
 
 export const parseXmlString = (xmlString) => {
   const parser = new xml2js.Parser({
@@ -154,3 +154,18 @@ export const doShortenerRedirect = (ctx) => {
 
   return false;
 };
+
+export class FetchError extends Error {
+  constructor(
+    public message: string = "",
+    public code: string,
+    public data: string
+  ) {
+    super();
+  }
+
+  toString() {
+    const suffix = this.data && ` Data: ${this.data.substring(0, 1000)}`;
+    return `Fetch: ${this.message}${suffix}`;
+  }
+}
